requirements.txt
attrs==21.2.0
betterproto==2.0.0b3
black==21.12b0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
debtcollector==1.22.0
factory-boy==3.2.1
Faker==9.8.0
firebase-admin==5.2.0
funcsigs==1.0.2
GeoAlchemy2==0.9.4
google-api-core==2.4.0
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpc-interceptor==0.13.0
grpcio==1.43.0
grpcio-status==1.43.0
grpcio-tools==1.43.0
grpclib==0.4.3rc1
h2==4.1.0
hpack==4.0.0
httplib2==0.20.2
hyperframe==6.0.1
idna==3.3
iniconfig==1.1.1
Jinja2==2.11.3
MarkupSafe==2.0.1
msgpack==1.0.3
multidict==5.2.0
mypy-extensions==0.4.3
oslodb-sqlalchemy-exceptions
packaging==21.2
pathspec==0.9.0
pg8000
pbr==5.7.0
platformdirs==2.4.0
pluggy==1.0.0
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.16.0
python-dateutil==2.8.2
python-decouple==3.5
requests==2.27.1
rsa==4.8
Rtree==0.9.7
Shapely==1.8.0
six==1.16.0
SQLAlchemy==1.4.26
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
